requirements.txt
Flask==2.1.0
PyGithub==1.55
python-dotenv==0.19.0
Flask-SSLify==0.1.5
gunicorn==20.1.0
openai==0.27.2
flask_session==0.4.0